[Add] bindings module
