+++ engine/CMakeLists.txt
@@ -30,6 +30,8 @@ target_add_module(Engine physics Physics)
 target_add_module(Engine steam Steam)
 target_add_module(Engine audio Audio)
 target_add_module(Engine user_interface UserInterface)
+target_add_module(Engine bindings Bindings)
 target_add_module(Engine ecs ECS)
+
 
 target_link_libraries(Engine INTERFACE OldEngine)
