+++ engine/bindings/public/wren_bindings.hpp
@@ -1,0 +1,4 @@
+#pragma once
+#include <wren_common.hpp>
+
+void BindEngineAPI(wren::ForeignModule& module);
