+++ engine/main.cpp
@@ -11,7 +11,7 @@
 #include "time_module.hpp"
 #include "ui_module.hpp"
 
-#include "utility/bind_math.hpp"
+#include "wren_bindings.hpp"
 
 int main(MAYBE_UNUSED int argc, MAYBE_UNUSED char* argv[])
 {
@@ -31,21 +31,10 @@ int main(MAYBE_UNUSED int argc, MAYBE_UNUSED char* argv[])
         .AddModule<ParticleModule>();
 
     auto& scripting = instance.GetModule<ScriptingModule>();
-    bindings::DefineMathTypes(scripting.GetForeignAPI());
 
-    // Add components here to expose them in scripting
-    {
-    }
-
-    // Add modules here to expose them in scripting
-    {
-        auto& engineAPI = scripting.GetEngineClass();
-        engineAPI.func<&WrenEngine::GetModule<TimeModule>>("GetTime");
-        engineAPI.func<&WrenEngine::GetModule<ECSModule>>("GetECS");
-    }
-
+    BindEngineAPI(scripting.GetForeignAPI());
     scripting.GenerateEngineBindingsFile();
-    instance.GetModule<ScriptingModule>().SetMainScript(instance, "game/game.wren");
+    scripting.SetMainScript(instance, "game/game.wren");
 
     return instance.Run();
 }
