--- engine/scripting/private/utility/bind_math.cpp
@@ -1,77 +1,0 @@
-#include "utility/bind_math.hpp"
-#include <glm/gtc/quaternion.hpp>
-
-namespace vector_ops
-{
-template <typename T>
-static T Default() { return {}; }
-
-template <typename T>
-static T Add(T& lhs, const T& rhs) { return lhs + rhs; }
-
-template <typename T>
-static T Sub(T& lhs, const T& rhs) { return lhs - rhs; }
-
-template <typename T>
-static T Neg(T& lhs) { return -lhs; }
-
-template <typename T>
-static T Mul(T& lhs, const T& rhs) { return lhs * rhs; }
-
-template <typename T>
-static bool Equals(T& lhs, const T& rhs) { return lhs == rhs; }
-
-template <typename T>
-static bool NotEquals(T& lhs, const T& rhs) { return lhs != rhs; }
-
-template <typename T>
-static T Normalized(T& v) { return glm::normalize(v); }
-
-template <typename T>
-static float Length(T& v) { return glm::length(v); }
-
-};
-
-template <typename T>
-void BindVectorTypeOperations(wren::ForeignKlassImpl<T>& klass)
-{
-    klass.template funcStaticExt<vector_ops::Default<T>>("Default");
-    klass.template funcExt<vector_ops::Add<T>>(wren::OPERATOR_ADD);
-    klass.template funcExt<vector_ops::Sub<T>>(wren::OPERATOR_SUB);
-    klass.template funcExt<vector_ops::Neg<T>>(wren::OPERATOR_NEG);
-    klass.template funcExt<vector_ops::Mul<T>>(wren::OPERATOR_MUL);
-    klass.template funcExt<vector_ops::Equals<T>>(wren::OPERATOR_EQUAL);
-    klass.template funcExt<vector_ops::NotEquals<T>>(wren::OPERATOR_NOT_EQUAL);
-    klass.template funcExt<vector_ops::Normalized<T>>("normalize");
-    klass.template funcExt<vector_ops::Length<T>>("length");
-}
-
-void bindings::DefineMathTypes(wren::ForeignModule& module)
-{
-    {
-        auto& vector2 = module.klass<glm::vec2>("Vec2");
-        vector2.ctor<float, float>();
-        vector2.var<&glm::vec2::x>("x");
-        vector2.var<&glm::vec2::y>("y");
-        BindVectorTypeOperations(vector2);
-    }
-
-    {
-        auto& vector3 = module.klass<glm::vec3>("Vec3");
-        vector3.ctor<float, float, float>();
-        vector3.var<&glm::vec3::x>("x");
-        vector3.var<&glm::vec3::y>("y");
-        vector3.var<&glm::vec3::z>("z");
-        BindVectorTypeOperations(vector3);
-    }
-
-    {
-        auto& quat = module.klass<glm::quat>("Quat");
-        quat.ctor<float, float, float, float>();
-        quat.var<&glm::quat::w>("w");
-        quat.var<&glm::quat::x>("x");
-        quat.var<&glm::quat::y>("y");
-        quat.var<&glm::quat::z>("z");
-        BindVectorTypeOperations(quat);
-    }
-}
