--- engine/scripting/public/utility/bind_math.hpp
@@ -1,7 +1,0 @@
-#pragma once
-#include "wren_common.hpp"
-
-namespace bindings
-{
-void DefineMathTypes(wren::ForeignModule& module);
-}
